Add files via upload
This dataset entails the property sales in the counties in Ireland in the year 2020 solely for the month of January. The dataset displays sales of apartments across all the  26 counties. The average house price was roughly around 256,000 Euros in 2020 (Slater, S. 2021). When it comes to the most expensive place to buy a house in Ireland, Dublin tops the chart. (Holland, L. Retrieved on 2021, June 25). The 26 counties have been divided into 4 Provinces – Connacht, Leinster, Munster, and Ulster (Republic of Ireland)  for better analysis of this dataset. I planned to analyze the most expensive and the least expensive counties in terms of property prices. Accordingly, one-sample t-tests and  2 sample t-tests would be conducted to compare the means of two groups, while conducting hypothesis testing.